1. Move signer tests to blockchain folder 2. Fix potensial swift memory leak

+++ src/Bitcoin/UnspentCalculator.h
@@ -11,7 +11,7 @@
 
 namespace TW::Bitcoin {
 
-using FeeCalculator = std::function<int64_t(size_t, size_t, int64_t)>;
+using FeeCalculator = std::function<int64_t(int64_t, int64_t, int64_t)>;
 using SingleInputFeeCalculator = std::function<int64_t(int64_t)>;
 
 class UnspentCalculator {
@@ -29,7 +29,7 @@ class UnspentCalculator {
         : calculate(std::move(calculateFee)), calculateSingleInput(std::move(calculateSingleInputFee)) {}
 
   private:
-    static int64_t calculateFee(size_t inputs, size_t outputs = 2, int64_t byteFee = 1);
+    static int64_t calculateFee(int64_t inputs, int64_t outputs = 2, int64_t byteFee = 1);
     static int64_t calculateSingleInputFee(int64_t byteFee);
 };
